1020550 - support 6.2 rhsm and return success even if auth fails

+++ spec/support/shared_examples/protected_action_shared_examples.rb
@@ -52,21 +52,23 @@ shared_examples_for "protected action" do
   end
   context "NOT ALLOWING me" do
     it "to it", :katello => true do #TODO headpin
-      if !defined?(on_success) && !defined?(before_success)
-        @controller.stub(action)
-        @controller.stub(:render)
-      end
-      session.delete(:current_organization_id)
-      login_user_by_described_class(unauthorized_user) if defined?  unauthorized_user
-      before_failure if defined?(before_failure)
-      if @controller.kind_of? Api::V1::ApiController
-        @controller.should_receive(:respond_for_exception).with { |e, options| options.try(:[], :status).should == :forbidden }
-      else
-        @controller.should_receive(:render_403)
-      end
+      if defined?  unauthorized_user
+        if !defined?(on_success) && !defined?(before_success)
+          @controller.stub(action)
+          @controller.stub(:render)
+        end
+        session.delete(:current_organization_id)
+        login_user_by_described_class(unauthorized_user)
+        before_failure if defined?(before_failure)
+        if @controller.kind_of? Api::V1::ApiController
+          @controller.should_receive(:respond_for_exception).with { |e, options| options.try(:[], :status).should == :forbidden }
+        else
+          @controller.should_receive(:render_403)
+        end
 
-      req
-      on_failure if defined?(on_failure)
+        req
+        on_failure if defined?(on_failure)
+      end
     end
   end
 end
